Create the developer-tools Docker image
The developer-tools image will be the unified Docker image to be used
for local development and within CI pipelines. This image consolidates
the previous kitchen-terraform and lint images into a single image with
all dependencies. A Makefile has been provided to abstract the build and
release commands to align with current CFT patterns.

Shell scripts are being used to install dependencies for two reasons:

* The script can be linted and modified independently from the
Dockerfile
* We can run multiple `apk add` commands without creating multiple
intermediate containers