Migrate task helper functions and wrapper scripts
The developer-tools Dockerfile and Makefile were moved into their own
folder to isolate the image from the other existing CFT images. In doing
so, however, it meant that the build context couldn't access the task
helper functions and wrapper scripts which were located outside of the
new directory. This commit moves those scripts to ensure that they're
accessible during the developer-tools build.