Provide scripts for each CI operation mode
Previously `setup_environment.sh` contained an if-statement to determine
whether the script was being sourced or executed directly. Because both
of those operation modes need to be supported, and to reduce complexity
and keep things clear, two separate scripts have been provided to
address each of those operation modes.

`/usr/local/bin/setup_environment.sh` requires SERVICE_ACCOUNT_JSON to
be set with the contents of a service account JSON string, and in turn
that script uses the SA data to set the required environment variables and
configuration files to ensure terraform, inspec, gsutil, and gcloud will
operate with those credentials. This script is intended for local
development where credentials need to be setup on an ongoing basis.

`/usr/local/bin/execute_with_credentials.sh` similarly requires
SERVICE_ACCOUNT_JSON and sources `/usr/local/bin/setup_environment.sh`
to setup the environment. In addition, any arguments provided to the
script will be executed with those credentials, and when the script
exits any credentials written to disk will be cleaned up. This script is
provided so individual commands can be executed in a one-off fashion
with the provided credentials.