Fix `unbound variable` bug
If `SERVICE_ACCOUNT_JSON` is not set and `execute_with_credentials.sh`
is executed, an `unbound variable` error is raised instead of getting
the error and advice on how to set `SERVICE_ACCOUNT_JSON`. See the
following:

    bash-5.0# unset SERVICE_ACCOUNT_JSON
    bash-5.0# /usr/local/bin/execute_with_credentials.sh gsutil acl get gs://glarizza-seed-bucket
    /usr/local/bin/setup_environment.sh: line 16: SERVICE_ACCOUNT_JSON: unbound variable

To fix this the conditional check that references `SERVICE_ACCOUNT_JSON`
contains an empty default for the variable which allows the error and
advice for setting the variable to be echoed out.