Use CMD instead of ENTRYPOINT
A previous commit (5e2901e) established `/bin/bash` as ENTRYPOINT, but
that logic was flawed because it both throws an error if you try and
specify a command inline to `docker run` and it's also not the best
use-case for ENTRYPOINT.

The base image for developer-tools is the ruby:2.6.3-alpine image, and
that image sets `CMD ["irb"]`. CMD is the correct command to use because
it can be overridden from the command line and it can also be overridden
by the developer-tools image. This commit changes ENTRYPOINT to CMD and
seems to give us what we need:

    $ docker run --rm -it cft/developer-tools:0.0.1
    bash-5.0# exit

    $ docker run --rm -it cft/developer-tools:0.0.1 bash
    bash-5.0# exit